AutomorphismGroup(poset)
Now uses graph representation to find the automorphism group on the flags. Before I modified the code for the function to be the default I ran this comparison on posetFE(3):

gap> p:=posetFE(3);
A poset using the IsPosetOfIndices representation
gap> AutomorphismGroup(p);time;
<permutation group with 3 generators>
1748
gap> AutomorphismGroup(p,true);time;
<permutation group with 4 generators>
82

So that's a 21 fold improvement in speed. WOW.